pyproject.toml
[project]
name = "git-ai"
version = "0.1.0"
description = "AI-powered GitHub PR description generator"
requires-python = ">=3.8"
dependencies = [
    "instructor[bedrock]",
    "python-dotenv>=1.0.0",
    "PyGithub>=2.1.1",
    "pydantic>=2.0.0",
    "boto3>=1.34.0",
]

[project.optional-dependencies]
dev = ["ruff==0.3.0", "pytest==8.0.0", "pytest-mock==3.14.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v"
markers = [
    "integration: marks tests that require external services (deselect with '-m \"not integration\"')",
]

[tool.ruff]
line-length = 88
target-version = "py38"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"
